cleanup weechat config
most "default" files and various extra scripts/plugins were removed in
favor of a small set of files needed for ui/color tweaks.

I'm not using weechat nearly as much as I used to, and it also doesn't
make sense to source control the "default" files that get automatically
created.